fix: :truck: Moved rotation to src/rotation/block.rs
Issue: #39
