+++ src/rotation/mod.rs
@@ -2,6 +2,7 @@ use bevy::prelude::*;
 
 use crate::AppState;
 
+mod block;
 mod events;
 
 #[derive(Event, Default)]
@@ -15,6 +16,7 @@ impl Plugin for RotationPlugin {
             .add_event::<BlockRotationEvent>()
             .add_systems(Update, (
                 events::rotation,
+                block::rotation,
             ).chain().run_if(in_state(AppState::Ingame)))
         ;
     }
